Move rustc-related logic from `RustToolchain` to `Rustc`
